Fixed #17; Fixed price bug and implement orders count by status in the main screen

+++ app/src/main/java/com/wcsm/confectionaryadmin/ui/view/CreateOrderScreen.kt
@@ -641,7 +641,7 @@ fun CreateOrderScreen(
                         if(value.isEmpty()) value = "0"
                         createOrderViewModel.updateCreateOrderState(
                             orderState.copy(
-                                price = value.toDouble()
+                                price = value
                             )
                         )
                         createOrderViewModel.createNewOrder()
